Add org name to main package name
This change renames the main package from `jseql-ffi` to `@cipherstash/jseql-ffi`.

This is consistent with other packages in our org. This change should also fix up the
release workflow since it's currently failing to publish the main package because
the NPM token is scoped to not allow packages scoped outside the org.