Add integration tests
This change wires up integration tests and adds steps for running
integration tests to CI.

The general approach is to use a separate package in
`./integration-tests` that installs `@cipherstash/protect-ffi` as
a local dependency. This keeps our TS config simple and allows
us to test against the build output (including types) instead
of only source files.

There's currently only a single test, but the intent is to add
more after we have the basic plumbing in place.