Add `clientOpts` to `newClient`
This change adds an optional `clientOpts` argument to `newClient`.

This value is a JSON-encoded string. The current options are (note that options are
snake_cased for consistency with the first arg, the encrypt schema):
- `client_key`
- `client_id`
- `access_key`
- `workspace_crn`

Also note that env vars and TOML files both take precedence over the `clientOpts`
arg (due to how precedence works in `cipherstash-client`).

Example usage:
```
newClient(
  JSON.stringify({tables: {}, v: 1}),
  JSON.stringify({
    workspace_crn: 'crn',
    access_key: 'access_key',
    client_id: 'client_id',
    client_key: 'client_key'
  })
)
```

This is admittedly ugly, but it gets the job done.
We should take a look at refactoring this after 1) the Neon version bump
and 2) refactoring config handling in `cipherstash-client`.